\name{Ospats-package}
\alias{Ospats-package}
\alias{Ospats}
\docType{package}
\title{
\packageTitle{Ospats}
}
\description{
\packageDescription{Ospats}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{Ospats}
\packageIndices{Ospats}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{Ospats}

Maintainer: \packageMaintainer{Ospats}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
