\name{Ospats-package}
\alias{Ospats-package}
\alias{Ospats}
\docType{package}
\title{
\packageTitle{Ospats}
}
\description{
\packageDescription{Ospats}
}
\details{
A bit more of a description of this package is required here
}

\author{
\packageAuthor{Ospats}

Maintainer: \packageMaintainer{Ospats}
}

\references{
\itemize{
\item Odgers, N.P., Sun, W., McBratney, A.B., Minasny, B., Clifford, D., (2014) \href{http://dx.doi.org/10.1016/j.geoderma.2013.09.024}{Disaggregating and harmonising soil map units through resampled classification trees}. Geoderma, 214-215: 91-100.
\item Odgers, N.P., Sun, McBratney, A.B., Minasny, B., (2015) \href{http://dx.doi.org/10.1016/j.geoderma.2014.09.009}{Digital soil property mapping and uncertainty estimation using soil class probability rasters}. Geoderma, 237:238 190-198.
}}


