\name{Ospats-package}
\alias{Ospats-package}
\alias{Ospats}
\docType{package}
\title{
\packageTitle{Ospats}
}
\description{
\packageDescription{Ospats}
}
\details{
Optimization of spatial stratification and allocation for stratified random sampling of points in the plane. Theory and discussion in both de Gruijter papers reference below.
}

\author{
\packageAuthor{Ospats}

Maintainer: \packageMaintainer{Ospats}
}

\references{
\itemize{
\item de Gruijter, J.J., McBratney, A.B., Minasny, B., Wheeler, I., Malone, B.P., Stockmann, U., (2016) \href{http://dx.doi.org/10.1016/j.geoderma.2015.11.010}{Farm-scale soil carbon auditing}. Geoderma 265, 120-130.
\item de Gruijter, J.J., Minasny, B., McBratney, A.B., (2015) \href{http://dx.doi.org/10.1093/jssam/smu024}{Optimizing Stratification and Allocation for Design-Based Estimation of Spatial Means Using Predictions with Error}. Journal of Survey Statistics and Methodology 3(1), 19-42.
}}


